pyproject.toml
[tool.poetry]
name = "rn-cli"
version = "0.2.4"
description = ""
authors = ["alex <alexander.smehnov@gmail.com>"]
readme = "README.md"
packages = [
    { include = "mb" },
]

[tool.poetry.dependencies]
python = "^3.8.1"
typer = "^0.9.0"
shellingham = "^1.5.4"
colorama = "^0.4.6"
requests = "^2.31.0"
pytest = "^8.0.0"
rich = "^13.7.0"
python-socketio = "^5.11.3"
readchar = "^4.1.0"


cryptography = "^43.0.0"
pylint = "^3.2.6"
textual = "^0.75.1"
base58 = "^2.1.1"
websockets = "^13.0"
pyyaml = "^6.0.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rn = 'mb.__main__:main'